rename embed.v1.js to embed.js
